multiprocessing package renamed multi_processing; grid_surface refactored into package; deprecated code removed
