+++ resqpy/__init__.py
@@ -13,7 +13,7 @@
     grid
     grid_surface
     lines
-    multiprocessing
+    multi_processing
     organize
     property
     rq_import
