+++ resqpy/grid/_grid_types.py
@@ -51,24 +51,18 @@ def is_regular_grid(grid_root):
     return grid_flavour(grid_root) == 'IjkBlockGrid'
 
 
-def any_grid(parent_model, uuid = None, grid_root = None, find_properties = True):
+def any_grid(parent_model, uuid = None, find_properties = True):
     """Returns a Grid or RegularGrid or UnstructuredGrid object depending on the extra metadata in the xml.
 
     arguments:
         parent_model (Model): the model within which the grid exists
         uuid (UUID): the uuid of the grid object to be instantiated
-        grid_root (lxml.Element, DEPRECATED): an alternative to passing the uuid
         find_properties (bool, default True): passed onward to the instantiation method
 
     note:
         full list of resqpy grid class objects which could be returned:
         Grid, RegularGrid, UnstructuredGrid, TetraGrid, HexaGrid, PyramidGrid, PrismGrid
     """
-
-    if grid_root is not None:
-        warnings.warn('Deprecated: use of grid_root arg to any_grid(); use uuid instead', DeprecationWarning)
-        if uuid is None:
-            uuid = rqet.uuid_for_part_root(grid_root)
 
     assert uuid is not None
