+++ resqpy/property/property_common.py
@@ -501,7 +501,6 @@ def selective_version_of_collection(
         collection,
         realization = None,
         support_uuid = None,
-        grid = None,  # for backward compatibility
         uuid = None,
         continuous = None,
         points = None,
@@ -526,7 +525,6 @@ def selective_version_of_collection(
        collection (PropertyCollection): an existing collection from which a subset will be returned as a new object
        realization (int, optional): realization number to filter on
        support_uuid (UUID or str, optional): UUID of supporting representation, to filter on
-       grid (Grid, DEPRECATED): for backward compatibility, use support_uuid instead
        uuid (UUID or str, optional): a property uuid to select a singleton property from the collection
        continuous (bool, optional): if True, continuous properties are selected; if False, discrete and categorical
        points (bool, optional): if True, points properties are selected; if False, they are excluded
@@ -561,15 +559,12 @@ def selective_version_of_collection(
        finally, the filters for all the attributes must be passed for a given member
        to be included in the returned collection; title is a synonym for the citation_title argument;
        related_uuid will pass if any relationship exists ('hard' or 'soft');
-       the grid keyword argument is maintained for backward compatibility: support_uuid argument takes precedence;
        the categorical boolean argument can be used to select only categorical (or non-categorical) properties,
        even though this is not explicitly held as a field in the internal dictionary
     """
 
     assert collection is not None
     view = rqp.PropertyCollection()
-    if support_uuid is None and grid is not None:
-        support_uuid = grid.uuid
     if support_uuid is not None:
         view.set_support(support_uuid = support_uuid, model = collection.model)
     if realization is not None:
